info/requirements.txt
arabic-reshaper==2.1.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
pika==1.1.0
Pillow==8.1.0
psycopg2-binary==2.8.6
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
reportlab==3.5.59
requests==2.24.0
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.11
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.5
zipp==3.4.0
gunicorn==19.9.0
